requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-MySQL==1.5.2
itsdangerous==2.0.0
Jinja2==3.0.0
mariadb==1.0.6
MarkupSafe==2.0.0
mysql-connector-python==8.0.25
protobuf==3.17.0
PyJWT==2.1.0
PyMySQL==1.0.2
six==1.16.0
Werkzeug==2.0.0
